[TFX-2240] core update and EODThresholdFutureValueRepository

+++ src/main/java/com/ihsmarkit/tfx/eod/service/EODCalculator.java
@@ -351,7 +351,7 @@ public class EODCalculator {
     }
 
     public Map<CurrencyPairEntity, List<BalanceTrade>> rebalanceLPPositions(final Stream<ParticipantPositionEntity> positions,
-                                                                            final Map<CurrencyPairEntity, BigDecimal> thresholds) {
+                                                                            final Map<CurrencyPairEntity, Long> thresholds) {
 
         return positions
             .map(tradeOrPositionMapper::convertPosition)
